+++ packages/performance-tests/test/TestContext.ts
@@ -8,27 +8,9 @@ import { Logger } from "@itwin/core-bentley";
 import { BriefcaseIdValue, IModelVersion } from "@itwin/core-common";
 import { AccessTokenAdapter, BackendIModelsAccess } from "@itwin/imodels-access-backend";
 import assert from "assert";
-import { generateTestIModel, IModelParams } from "./iModelUtils";
+import { generateTestIModel } from "./iModelUtils";
 
 const loggerCategory = "TestContext";
-
-export interface ReporterInfo {
-  /* eslint-disable @typescript-eslint/naming-convention */
-  "Id": string;
-  "T-shirt size": string;
-  "Gb size": string;
-  "Branch Name": string;
-  "Federation Guid Saturation": number;
-  /* eslint-enable @typescript-eslint/naming-convention */
-}
-
-export interface ReporterEntry {
-  testSuite: string;
-  testName: string;
-  valueDescription: string;
-  value: number;
-  info?: ReporterInfo;
-}
 
 export interface BriefcaseArgs {
   fileName: string;
@@ -80,16 +62,8 @@ export async function *getTestIModels(filter: (iModel: TestIModel) => boolean) {
       }
     }
   }
-  const iModelParamsFedguids: IModelParams = {
-    numElements: 100000,
-    fedGuids: true,
-  };
-  yield generateTestIModel(iModelParamsFedguids);
-  const iModelParamsNoFedguids: IModelParams = {
-    numElements: 100000,
-    fedGuids: false,
-  };
-  yield generateTestIModel(iModelParamsNoFedguids);
+  yield generateTestIModel({ numElements: 100_000, fedGuids: true });
+  yield generateTestIModel({ numElements: 100_000, fedGuids: false });
 }
 
 export async function downloadAndOpenBriefcase(briefcaseArg: Omit<RequestNewBriefcaseArg, "accessToken">): Promise<BriefcaseDb> {
